refactor: common-runtime -> internal/projectinit
