+++ internal/projectinit/scaffolding/bin/.ftl@latest.pkg
