+++ internal/projectinit/scaffolding/bin/README.hermit.md
