+++ internal/projectinit/scaffolding/bin/activate-hermit
