+++ internal/projectinit/scaffolding/bin/ftl-controller
