+++ internal/projectinit/scaffolding/go.mod
